% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseqPlants-package.R
\docType{package}
\name{scRNAseqPlants-package}
\alias{scRNAseqPlants-package}
\alias{scRNAseqPlants}
\title{Title}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://support.bioconductor.org/t/scRNAseqPlants}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
